fixing the autocert _get_domains function call
msg from DigiCert:

Scott,

If you modify your URL to remove ?container_id=76790 then it will return the list of domains.

https://www.digicert.com:443/services/v2/domain

Are you trying to report on which domains are allowed for a particular division? In order to find the list of domains allowed for a division you can use the “view a container” endpoint.

Documentation for the “view a container” endpoint:
https://www.digicert.com/services/v2/documentation/container/container

Endpoint URL:
https://www.digicert.com/services/v2/container/{container_id}

(allowed_domain_names is part of the data returned by that endpoint)

Thank you,

Darin Andrew, CISSP